backend/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
blinker==1.6.3
blis==0.7.11
catalogue==2.0.10
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
cloudpathlib==0.16.0
cohere==4.32
confection==0.1.4
cymem==2.0.8
distro==1.8.0
dnspython==2.4.2
exceptiongroup==1.2.0
fastavro==1.8.2
Flask==3.0.0
Flask-Cors==4.0.0
frozenlist==1.4.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
langcodes==3.3.0
llmsherpa==0.1.3
MarkupSafe==2.1.3
multidict==6.0.4
murmurhash==1.0.10
numpy==1.26.2
openai==1.3.7
packaging==23.2
preshed==3.0.9
pydantic==2.5.2
pydantic_core==2.14.5
pymongo==4.5.0
python-dotenv==1.0.0
requests==2.31.0
smart-open==6.4.0
sniffio==1.3.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
thinc==8.2.2
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.8.0
urllib3==2.0.7
wasabi==1.1.2
weasel==0.3.4
Werkzeug==3.0.1
yarl==1.9.2
zipp==3.17.0
